+++ backend/todo/constants/constants.py
@@ -1,0 +1,28 @@
+# todo/constants.py
+
+# Choices for Task priority
+PRIORITY_CHOICES = [
+    ("High", "High"),
+    ("Medium", "Medium"),
+    ("Low", "Low"),
+]
+
+# Choices for Task status
+STATUS_CHOICES = [
+    ("Pending", "Pending"),
+    ("Started", "Started"),
+    ("Completed", "Completed"),
+]
+
+# Choices for Task category
+CATEGORY_CHOICES = [
+    ("Frontend", "Frontend"),
+    ("Backend", "Backend"),
+]
+
+# Choices for context entry source
+CONTEXT_SOURCE_CHOICES = [
+    ('whatsapp', 'WhatsApp'),
+    ('email', 'Email'),
+    ('note', 'Note'),
+]
